+++ ArtBack.Domain/Entities/Vendor.cs
@@ -4,7 +4,8 @@ public class Vendor: Entity
 {
     public required int ArtworkCount { get; set; }
     public required int SoldArtworkCount { get; set; }
-    public required double Profits { get; set; }
+    public required decimal Profits { get; set; }
+    
     public ICollection<Artwork>? Artworks { get; set; }
     
 }
